♻️(front) create a dedicated component to manage document title form
Maintaining the title state in the DashboardDocument was not easy and
has unwanted behaviour like updating the form input once already render.
We choose to create a dedicated component, rendered only when the
document is in ready state. The document's title should not be updated
anymore at this step and we should not have unwanted behaviour anymore.
